// apps/jobs/web/src/lib/api/generated/model/candidateProfile.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateProfileUser } from './candidateProfileUser';
import type { CandidateProfileVisibility } from './candidateProfileVisibility';
import type { CandidateProfileDesiredFunctionsItem } from './candidateProfileDesiredFunctionsItem';
import type { CandidateProfileWorkplacePrefsItem } from './candidateProfileWorkplacePrefsItem';
import type { CandidateProfileExperienceItem } from './candidateProfileExperienceItem';
import type { CandidateProfileEducationItem } from './candidateProfileEducationItem';

export interface CandidateProfile {
  user: CandidateProfileUser;
  /** @nullable */
  headline: string | null;
  /** @nullable */
  location: string | null;
  openToWork: boolean;
  visibility: CandidateProfileVisibility;
  desiredFunctions: CandidateProfileDesiredFunctionsItem[];
  desiredLocations: string[];
  workplacePrefs: CandidateProfileWorkplacePrefsItem[];
  experience: CandidateProfileExperienceItem[];
  education: CandidateProfileEducationItem[];
  skills: string[];
  /** @nullable */
  portfolioUrl: string | null;
  links: string[];
}