// apps/jobs/web/src/lib/api/generated/model/candidateProfileInput.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateProfileInputVisibility } from './candidateProfileInputVisibility';
import type { CandidateProfileInputDesiredFunctionsItem } from './candidateProfileInputDesiredFunctionsItem';
import type { CandidateProfileInputWorkplacePrefsItem } from './candidateProfileInputWorkplacePrefsItem';
import type { CandidateProfileInputExperienceItem } from './candidateProfileInputExperienceItem';
import type { CandidateProfileInputEducationItem } from './candidateProfileInputEducationItem';

export interface CandidateProfileInput {
  /** @maxLength 120 */
  headline?: string;
  /** @maxLength 80 */
  location?: string;
  openToWork?: boolean;
  visibility?: CandidateProfileInputVisibility;
  /** @maxItems 6 */
  desiredFunctions?: CandidateProfileInputDesiredFunctionsItem[];
  /** @maxItems 6 */
  desiredLocations?: string[];
  /** @maxItems 3 */
  workplacePrefs?: CandidateProfileInputWorkplacePrefsItem[];
  /** @maxItems 20 */
  experience?: CandidateProfileInputExperienceItem[];
  /** @maxItems 10 */
  education?: CandidateProfileInputEducationItem[];
  /** @maxItems 30 */
  skills?: string[];
  /** @maxLength 500 */
  portfolioUrl?: string;
  /** @maxItems 6 */
  links?: string[];
}