// apps/jobs/web/src/lib/api/generated/model/candidateProfileInputVisibility.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */

export type CandidateProfileInputVisibility = typeof CandidateProfileInputVisibility[keyof typeof CandidateProfileInputVisibility];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CandidateProfileInputVisibility = {
  public: 'public',
  recruiters: 'recruiters',
  private: 'private',
} as const;